pyproject.toml
[tool.poetry]
name = "slack_bot"
version = "0.1.1"
description = "Data team slack bot"
authors = ["Hyperskill Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
slack-bolt = "1.18.0"
python-dotenv = "1.0.0"
pygithub = "1.59.1"
structlog = "23.1.0"
structlog-sentry = "2.0.3"
tiktoken = "0.5.1"
openai = "1.3.0"
trafilatura = "1.6.2"
clickhouse_driver = "0.2.6"
pandas = "2.1.3"
infi-clickhouse-orm = "2.1.3"

[tool.poetry.group.dev.dependencies]
black = "23.9.1"
mypy = "1.5.1"
ruff = "0.0.290"
types-requests = "2.31.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = [
    "A",
    "ANN",
    "B",
    "BLE",
    "C4",
    "C90",
    "COM",
    "D",
    "DJ",
    "DTZ",
    "E",
    "ERA",
    "EXE",
    "F",
    "G",
    "I",
    "ICN",
    "INP",
    "INT",
    "ISC",
    "N",
    "NPY",
    "PD",
    "PGH",
    "PIE",
    "PL",
    "PT",
    "PTH",
    "PYI",
    "Q",
    "RET",
    "RSE",
    "RUF",
    "S",
    "SIM",
    "SLF",
    "T10",
    "T20",
    "TCH",
    "TID",
    "TRY",
    "UP",
    "W",
    "YTT",
]
ignore = [
    "ANN002",
    "ANN003",
    "ANN101",
    "ANN102",
    "ANN401",
    "B019",
    "COM812",
    "D100",
    "D101",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "EXE002",
    "N806",
    "S101",
    "SIM108",
    "TRY003",
]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
line-length = 88
target-version = "py311"

[tool.ruff.mccabe]
max-complexity = 21

[tool.ruff.pydocstyle]
convention = "google"

[tool.ruff.pep8-naming]
classmethod-decorators = [
    "classmethod",
]

[tool.ruff.pylint]
max-args = 5
max-branches = 25
max-returns = 7
max-statements = 54

[tool.ruff.isort]
combine-as-imports = true
known-third-party = ["elasticsearch"]
order-by-type = false
required-imports = ["from __future__ import annotations"]

[tool.black]
line-length = 88
target-version = ['py311']
skip-string-normalization = true

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_any_generics = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
explicit_package_bases = true
ignore_errors = false
ignore_missing_imports = true
implicit_reexport = true
strict_equality = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true
